Move scrub test to a separate file
